feat: подготовить проект
